refactor(name) ladder에서 name 분리
