+++ src/test/java/nextstep/step2/LadderTest.java
@@ -1,24 +1,17 @@
 package nextstep.step2;
 
+import nextstep.step2.domain.Ladder;
+import nextstep.step2.domain.Line;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
 import java.util.List;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
 
 @DisplayName("사다리 클래스 테스트")
 public class LadderTest {
-
-    @Test
-    @DisplayName("이름이 5자가 넘으면 에러가 나는 테스트")
-    void validateNameLengthTest() {
-        String names = "pobi12,honux1,crong1,jk1234";
-        int inputLadderHeight = 3;
-        assertThatIllegalArgumentException().isThrownBy(() -> new Ladder(names, inputLadderHeight));
-    }
 
     @ParameterizedTest
     @ValueSource(ints = {3, 4, 5, 6})
@@ -26,7 +19,7 @@ public class LadderTest {
     void inputNamesAndLadderHeightTest(int inputLadderHeight) {
         String names = "pobi,honux,crong,jk";
         Ladder ladder = new Ladder(names, inputLadderHeight);
-        int joinedPeopleCount = ladder.getNames().size();
+        int joinedPeopleCount = ladder.getNames().getParticipantsCount();
         List<Line> ladders = ladder.getLadders();
         assertThat(ladders).hasSize(inputLadderHeight);
         ladders.forEach(line -> assertThat(line.getOneLayerLadder()).hasSize(joinedPeopleCount));
